+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/abc/avm2/parser/script/ConstAVM2Item.java
@@ -16,10 +16,16 @@
  */
 package com.jpexs.decompiler.flash.abc.avm2.parser.script;
 
+import com.jpexs.decompiler.flash.SourceGeneratorLocalData;
+import com.jpexs.decompiler.flash.abc.avm2.instructions.AVM2Instructions;
 import com.jpexs.decompiler.flash.abc.avm2.model.AVM2Item;
 import com.jpexs.decompiler.flash.helpers.GraphTextWriter;
+import com.jpexs.decompiler.graph.CompilationException;
+import com.jpexs.decompiler.graph.GraphSourceItem;
 import com.jpexs.decompiler.graph.GraphTargetItem;
+import com.jpexs.decompiler.graph.SourceGenerator;
 import com.jpexs.decompiler.graph.model.LocalData;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
@@ -72,6 +78,7 @@ public class ConstAVM2Item extends AVM2Item {
 
     /**
      * Check if is static
+     *
      * @return Is static
      */
     public boolean isStatic() {
@@ -80,6 +87,7 @@ public class ConstAVM2Item extends AVM2Item {
 
     /**
      * Constructor.
+     *
      * @param metadata Metadata
      * @param pkg Package
      * @param customNamespace Custom namespace
@@ -110,11 +118,29 @@ public class ConstAVM2Item extends AVM2Item {
 
     @Override
     public GraphTargetItem returnType() {
-        return type;
+        return null;
     }
 
     @Override
     public boolean hasReturnValue() {
-        return true;
+        return false;
+    }
+
+    @Override
+    public List<GraphSourceItem> toSource(SourceGeneratorLocalData localData, SourceGenerator generator) throws CompilationException {
+        AVM2SourceGenerator agen = (AVM2SourceGenerator) generator;
+        int ns = agen.genNs(localData.importedClasses, pkg.name, pkg, localData.openedNamespaces, localData, line);
+        if (type.toString().equals("Namespace")) {
+            return new ArrayList<>();
+        }
+
+        List<GraphSourceItem> ret = new ArrayList<>();
+        if (value != null) {
+            ret.add(ins(AVM2Instructions.FindProperty, agen.traitName(ns, var)));
+            localData.isStatic = true;
+            ret.addAll(agen.toInsList(value.toSource(localData, agen)));
+            ret.add(ins(AVM2Instructions.InitProperty, agen.traitName(ns, var)));
+        }
+        return ret;
     }
 }